+++ apps/auth-service/src/routes/auth/utils/decodeJWT.ts
@@ -1,0 +1,16 @@
+import * as jose from "jose";
+
+type GovIdJwtPayload = {
+  surname: string;
+  givenName: string;
+  email: string;
+};
+
+export default (token: string) => {
+  const decoded = jose.decodeJwt<jose.JWTPayload & GovIdJwtPayload>(token);
+  return {
+    firstName: decoded.givenName,
+    lastName: decoded.surname,
+    email: decoded.email,
+  };
+};
