+++ apps/auth-service/src/routes/index.ts
@@ -1,20 +1,10 @@
-// import * as url from "url";
 import { FastifyInstance } from "fastify";
-// import path from "path";
 import auth from "./auth/index.js";
 import healthCheck from "./healthcheck.js";
+import users from "./users/index.js";
 
-// const __dirname = url.fileURLToPath(new URL(".", import.meta.url));
 export default async function routes(app: FastifyInstance) {
   app.register(healthCheck);
-
-  // app.register(import("@fastify/static"), {
-  //   root: path.join(__dirname, "static"),
-  //   prefix: "/",
-  //   index: "index.html",
-  //   list: false,
-  //   constraints: {},
-  // });
-
   app.register(auth, { prefix: "/auth" });
+  app.register(users, { prefix: "/users" });
 }
