+++ apps/auth-service/src/types/index.d.ts
@@ -20,5 +20,6 @@ declare module "fastify" {
     Logger = FastifyLoggerInstance,
   > {
     verifyUser: (request: FastifyRequest, reply: FastifyReply) => Promise<void>;
+    config: { [key: string]: string };
   }
 }
