+++ apps/timeline-api/app.ts
@@ -66,31 +66,6 @@ export async function build(opts?: FastifyServerOptions) {
     database: process.env.POSTGRES_DB_NAME,
   });
 
-  app.setErrorHandler((error, _request, reply) => {
-    app.log.error(error);
-    if (error instanceof Error && error.name !== "error") {
-      reply
-        .code(error.statusCode || 500)
-        .type("application/json")
-        .send({
-          message: error.message,
-          error,
-          code: error.code || "INTERNAL_SERVER_ERROR",
-          statusCode: error.statusCode || 500,
-          time: new Date().toISOString(),
-        });
-      return;
-    }
-
-    reply.code(500).type("application/json").send({
-      message: error.message,
-      error: "Internal Server Error",
-      code: "INTERNAL_SERVER_ERROR",
-      statusCode: 500,
-      time: new Date().toISOString(),
-    });
-  });
-
   app.register(healthCheck);
 
   app.register(routes, { prefix: "/api/v1" });
